updated memove datatyp[e

+++ ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: jmanani <jmanani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 16:06:45 by jmanani           #+#    #+#             */
-/*   Updated: 2025/10/15 14:44:52 by jmanani          ###   ########.fr       */
+/*   Updated: 2025/10/15 15:45:30 by jmanani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dest, const void *src, int n)
+#include "libft.h"
+
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	char	*temp_src;
 	char	*temp_dest;
